Move `JsMinifyOptions`
